requirements.txt
flask==2.1.1
pandas==1.4.1
numpy==1.22.3
matplotlib==3.5.1
python-dotenv==0.20.0
flask-sqlalchemy==2.5.1
flask-login==0.6.0
beautifulsoup4==4.11.1
requests==2.27.1
tabula-py==2.3.0
